+++ src/router/index.tsx
@@ -1,0 +1,9 @@
+import { createBrowserRouter } from 'react-router-dom';
+import MainPage from '../pages/MainPage';
+
+export const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <MainPage />,
+  },
+]);
